refactor: starting game page to separate component
