+++ src/components/ConstructorGame/Start-screen/Start-screen.module.css
@@ -1,0 +1,61 @@
+.my-game {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background: #fdebff;
+  border-radius: 50px;
+  max-width: 1022px;
+  width: 80%;
+  min-height: 560px;
+  text-align: center;
+  padding: 5px;
+}
+
+.title {
+  font-weight: 900;
+  font-size: 4.8rem;
+  line-height: 5.6rem;
+  text-align: center;
+  text-transform: uppercase;
+  margin-bottom: 0.2rem;
+
+  color: #733999;
+}
+
+.text {
+  margin: 20px 0;
+  font-size: 1.8rem;
+  line-height: 2.1rem;
+  text-align: center;
+
+  color: rgba(115, 57, 153, 0.8);
+}
+
+.cat_sleeping {
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  width: 210px;
+  height: 142px;
+}
+
+.play-button {
+  display: flex;
+  margin-top: 3.2rem;
+  margin-bottom: 20px;
+  width: 60px;
+  height: 60px;
+  background-color: inherit;
+  border: none;
+}
+
+.play-button:hover {
+  cursor: pointer;
+}
+
+.play {
+  width: 60px;
+  height: 60px;
+}
