+++ src/utils/shuffle.ts
@@ -1,0 +1,8 @@
+export const shuffle = (array: any) => {
+  const arrCopy = [...array];
+  for (let i = arrCopy.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [arrCopy[i], arrCopy[j]] = [arrCopy[j], arrCopy[i]];
+  }
+  return arrCopy;
+};
